Added timetravel functionality to world overview
Just added a GtkScale in world overview to be able to change the time
showed in clocks from -12h to +12h so you can view what time is in
other part of the world not only at current time.

This is useful when you need to find a meeting date or something with
someone who lives in other timezone, so you can view in real time what
time will be in your local timezone and in the other person timezone.